feat: refactor sidebar navigation, enhance app header with suspense, and remove demo page
