+++ apps/web/components/navbar/sidebar-config.ts
@@ -54,7 +54,7 @@ export const data = {
     {
       title: "Overview",
       url: "/overview",
-      anotherUrl: "/demo",
+      anotherUrl: "/",
       icon: LayoutDashboard,
     },
     {
